requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp_socks==0.10.1
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.4
click==8.2.1
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.1
flask-pydantic-api==1.7.0
frozenlist==1.7.0
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.4
packaging==25.0
pillow==12.0.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-socks==2.7.2
requests==2.32.5
SQLAlchemy==2.0.43
tomli==2.3.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
